+++ server/mock/store.go
@@ -8,7 +8,6 @@ import (
 	dbm "github.com/cosmos/cosmos-db"
 	protoio "github.com/cosmos/gogoproto/io"
 
-	"cosmossdk.io/store/metrics"
 	pruningtypes "cosmossdk.io/store/pruning/types"
 	snapshottypes "cosmossdk.io/store/snapshots/types"
 	storetypes "cosmossdk.io/store/types"
@@ -53,10 +52,6 @@ func (ms multiStore) SetTracer(w io.Writer) storetypes.MultiStore {
 }
 
 func (ms multiStore) AddListeners(keys []storetypes.StoreKey) {
-	panic("not implemented")
-}
-
-func (ms multiStore) SetMetrics(metrics.StoreMetrics) {
 	panic("not implemented")
 }